feat: Expose real time dispatch changes function
When in a React environment, we use the RealTimeQueries component to
subscribe to a doctype changes and keep the internal store updated.

To be able to do the same thing manually for non React environment,
we expose the underlaying functions : dispatchCreate, dispatchCreate,
dispatchDelete.
