+++ packages/cozy-client/src/store/index.js
@@ -117,3 +117,5 @@ export {
   receiveMutationResult,
   receiveMutationError
 } from './mutations'
+
+export { dispatchCreate, dispatchUpdate, dispatchDelete } from './realtimes'
