+++ packages/cozy-client/types/store/realtimes.d.ts
@@ -1,0 +1,4 @@
+export function dispatchCreate(client: CozyClient, doctype: import("../types").Doctype, couchDBDoc: import("../types").CouchDBDocument): void;
+export function dispatchUpdate(client: CozyClient, doctype: import("../types").Doctype, couchDBDoc: import("../types").CouchDBDocument): void;
+export function dispatchDelete(client: CozyClient, doctype: import("../types").Doctype, couchDBDoc: import("../types").CouchDBDocument): void;
+import CozyClient from "../CozyClient";
